Make sure the custom trigger component checks that the trigger is an event trigger
